fix rocket rendering

+++ client/src/commandexecuter.h
@@ -27,7 +27,7 @@ class CommandExecuter : public Thread {
   void run();
  private:
   void renderMovingSprite(double x, double y, uint32_t itemId);
-  void loadNewTexture(uint32_t x, uint32_t y, uint32_t yamlId, uint32_t uniqueId);
+  void loadNewTexture(double x, double y, uint32_t yamlId, uint32_t uniqueId);
   void removeSpriteWithId(int itemId);
   void playShootingSounds(int shooterId);
   void playDoorOpeningSound(int x, int y);
